+++ src/rinex/index.ts
@@ -4,3 +4,4 @@ export * from './parser';
 export * from './warnings';
 export * from './nav';
 export * from './nav-writer';
+export * from './ionex';